# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "customer@prepareurself"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Register a user to prepareurself",
                "operationId": "App\\Http\\Controllers\\UserAuthController::register",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "first name of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "last name of user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password of user Min length 8 characters",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{[error_code=>1,msg=>'Invalid user data'],[error_code=>0,msg=>'Registeration Successfully Done']}"
                    }
                }
            }
        },
        "/api/check-username": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Register a user to prepareurself",
                "operationId": "App\\Http\\Controllers\\UserAuthController::checkUserName",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "check unique username of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{[error_code=>2,msg=>'Username Already Exits'],[error_code=>0,msg=>'Username available'],[error_code=>1,msg=>'Missing user name']}"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "login a user to prepareurself",
                "operationId": "App\\Http\\Controllers\\UserAuthController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{[error_code=>0,msg=>'login Successfully Done']}"
                    }
                }
            }
        },
        "/api/update-user": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "login a user to prepareurself",
                "operationId": "App\\Http\\Controllers\\UserAuthController::updateUserData",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "first name of user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "last name of user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dob",
                        "in": "query",
                        "description": "date of birth of user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "user phone number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "android_token",
                        "in": "query",
                        "description": "token for notification",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "preferences[]",
                        "in": "query",
                        "description": "User Preferences Please enter only integers",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{[error_code=>0,msg=>'Update Successfully Done']}"
                    }
                }
            }
        },
        "/api/update-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "update a user password in prepareurself",
                "operationId": "App\\Http\\Controllers\\UserAuthController::updatePassword",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "oldPassword",
                        "in": "query",
                        "description": "Old password of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "newPassword",
                        "in": "query",
                        "description": "New Password of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{[error_code=>0,msg=>'Successfully Password Changed']}"
                    }
                }
            }
        },
        "/api/forget-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Send Forgot Password Mail to user in prepareurself",
                "operationId": "App\\Http\\Controllers\\UserAuthController::forgetPassword",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{[error_code=>0,msg=>'Forgot password Mail sent to your email']}"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key_security_example": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "api_key",
                "in": "header"
            }
        }
    }
}